Conflicts:
	account-management-app/src/main/java/org/duracloud/account/app/controller/AccountController.java
	account-management-app/src/main/java/org/duracloud/account/app/controller/FullAccountForm.java
	account-management-app/src/main/java/org/duracloud/account/app/controller/NewAccountForm.java
	account-management-app/src/main/java/org/duracloud/account/app/controller/ServiceRepoForm.java
	account-management-app/src/main/java/org/duracloud/account/app/controller/ServiceRepositoryController.java
	account-management-app/src/main/java/org/duracloud/account/flow/createaccount/CreateAccountAction.java
	account-management-app/src/main/webapp/WEB-INF/config/security-config.xml
	account-management-app/src/test/java/org/duracloud/account/app/controller/AccountControllerTest.java
	account-management-app/src/test/java/org/duracloud/account/app/controller/AccountsControllerTest.java
	account-management-app/src/test/java/org/duracloud/account/app/controller/AmaTestBase.java
	account-management-app/src/test/java/org/duracloud/account/app/controller/ServiceRepositoryControllerTest.java
	account-management-common/src/main/java/org/duracloud/account/common/domain/ComputeProviderAccount.java
	account-management-common/src/main/java/org/duracloud/account/common/domain/ServerDetails.java
	account-management-compute-amazon/src/main/java/org/duracloud/account/compute/AmazonComputeProvider.java
	account-management-db-amazon/src/main/java/org/duracloud/account/db/amazonsimple/converter/DuracloudServerDetailsConverter.java
	account-management-db-amazon/src/test/java/org/duracloud/account/db/amazonsimple/TestDuracloudServerDetailsRepoImpl.java
	account-management-db-amazon/src/test/java/org/duracloud/account/db/amazonsimple/converter/DuracloudServerDetailsConverterTest.java
	account-management-db-hybrid/src/main/java/org/duracloud/account/db/hybrid/HybridDBRepoMgr.java
	account-management-db-hybrid/src/test/java/org/duracloud/account/db/hybrid/HybridDBRepoMgrTest.java
	account-management-db-model/src/main/java/org/duracloud/account/db/model/ServicePlan.java
	account-management-db-util/src/main/java/org/duracloud/account/db/util/DbUtil.java
	account-management-db/src/main/java/org/duracloud/account/db/DuracloudRepoMgr.java
	account-management-db/src/main/java/org/duracloud/account/db/IdUtil.java
	account-management-db/src/main/java/org/duracloud/account/db/impl/IdUtilImpl.java
	account-management-db/src/test/java/org/duracloud/account/db/impl/IdUtilImplTest.java
	account-management-init/src/test/java/org/duracloud/account/init/InitializerTest.java
	account-management-monitor/src/test/java/org/duracloud/account/monitor/hadoop/HadoopServiceMonitorTest.java
	account-management-util/pom.xml
	account-management-util/src/main/java/org/duracloud/account/db/util/RootAccountManagerService.java
	account-management-util/src/main/java/org/duracloud/account/db/util/error/DuracloudServiceRepositoryNotAvailableException.java
	account-management-util/src/main/java/org/duracloud/account/db/util/impl/DuracloudInstanceServiceImpl.java
	account-management-util/src/main/java/org/duracloud/account/db/util/instance/DurabossUpdater.java
	account-management-util/src/main/java/org/duracloud/account/db/util/instance/impl/DurabossUpdaterImpl.java
	account-management-util/src/main/java/org/duracloud/account/db/util/instance/impl/InstanceConfigUtilImpl.java
	account-management-util/src/main/java/org/duracloud/account/db/util/instance/impl/InstanceUpdaterImpl.java
	account-management-util/src/main/java/org/duracloud/account/db/util/security/SecurityContextUtil.java
	account-management-util/src/main/java/org/duracloud/account/util/impl/AccountManagerServiceImpl.java
	account-management-util/src/main/java/org/duracloud/account/util/impl/DuracloudProviderAccountUtil.java
	account-management-util/src/main/java/org/duracloud/account/util/impl/RootAccountManagerServiceImpl.java
	account-management-util/src/test/java/org/duracloud/account/util/impl/AccountServiceImplTest.java
	account-management-util/src/test/java/org/duracloud/account/util/impl/AccountServiceSecuredImplTest.java
	account-management-util/src/test/java/org/duracloud/account/util/impl/DuracloudInstanceManagerServiceImplTest.java
	account-management-util/src/test/java/org/duracloud/account/util/impl/DuracloudInstanceServiceImplTest.java
	account-management-util/src/test/java/org/duracloud/account/util/impl/DuracloudServiceTestBase.java
	account-management-util/src/test/java/org/duracloud/account/util/impl/RootAccountManagerServiceImplTest.java
	account-management-util/src/test/java/org/duracloud/account/util/instance/impl/DurabossUpdaterImplTest.java
	account-management-util/src/test/java/org/duracloud/account/util/instance/impl/InstanceConfigUtilImplTest.java
	account-management-util/src/test/java/org/duracloud/account/util/sys/impl/EventMonitorTestBase.java
	pom.xml
	shib-idp2/pom.xml

+++ account-management-app/src/main/java/org/duracloud/account/app/controller/AbstractController.java
@@ -3,25 +3,19 @@
  */
 package org.duracloud.account.app.controller;
 
-import java.util.Set;
-
-import org.duracloud.account.common.domain.DuracloudUser;
-import org.duracloud.account.common.domain.Role;
-import org.duracloud.account.util.DuracloudUserService;
+import org.duracloud.account.db.model.DuracloudUser;
+import org.duracloud.account.db.model.Role;
+import org.duracloud.account.db.util.DuracloudUserService;
 import org.duracloud.account.util.UserFeedbackUtil;
-import org.duracloud.common.error.DuraCloudRuntimeException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContext;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.springframework.web.servlet.view.RedirectView;
+
+import java.util.Set;
 
 /**
  * Base class for all controllers.
@@ -68,7 +62,7 @@ public abstract class AbstractController {
     }
 
     protected void setUserRights(DuracloudUserService userService,
-                                 int accountId, int userId, Role role) {
+                                 Long accountId, Long userId, Role role) {
         Set<Role> roles = role.getRoleHierarchy();
 
         userService.setUserRights(accountId,
@@ -81,7 +75,7 @@ public abstract class AbstractController {
      * @return
      */
     protected boolean accountHasMoreThanOneOwner(
-        Set<DuracloudUser> users, int accountId) {
+        Set<DuracloudUser> users, Long accountId) {
         int ownerCount = 0;
         for (DuracloudUser u : users) {
             if (u.isOwnerForAcct(accountId) && !u.isRootForAcct(accountId)) {
