+++ account-management-app/src/main/java/org/duracloud/account/util/UrlHelper.java
@@ -1,11 +1,11 @@
 package org.duracloud.account.util;
 
 public class UrlHelper {
-    public static String formatAccountId(int accountId, String pattern){
+    public static String formatAccountId(Long accountId, String pattern){
         return pattern.replace("{accountId}", String.valueOf(accountId));
     }
 
-    public static String formatId(int accountId, String pattern) {
+    public static String formatId(Long accountId, String pattern) {
         return pattern.replace("{id}", String.valueOf(accountId));
     }
 }
