+++ account-management-app/src/test/java/org/duracloud/account/app/controller/ServerImageControllerTest.java
@@ -3,7 +3,7 @@
  */
 package org.duracloud.account.app.controller;
 
-import org.duracloud.account.util.RootAccountManagerService;
+import org.duracloud.account.db.util.RootAccountManagerService;
 import org.easymock.EasyMock;
 import org.junit.Before;
 import org.junit.Test;
@@ -25,7 +25,7 @@ public class ServerImageControllerTest extends AmaControllerTestBase {
     public void testAdd() throws Exception {
         // set up mocks, and args
         ServerImageForm serverImageForm = new ServerImageForm();
-        serverImageForm.setProviderAccountId(1);
+        serverImageForm.setProviderAccountId(1L);
         serverImageForm.setProviderImageId("id");
         serverImageForm.setDescription("desc");
         serverImageForm.setVersion("version");
@@ -49,20 +49,20 @@ public class ServerImageControllerTest extends AmaControllerTestBase {
 
     @Test
     public void testDelete() throws Exception {
-        rootAccountManagerService.deleteServerImage(1);
+        rootAccountManagerService.deleteServerImage(1L);
         EasyMock.expectLastCall();
 
         replayMocks();
-        this.serverImagesController.delete(1, redirectAttributes);
+        this.serverImagesController.delete(1L, redirectAttributes);
     }
 
     @Test
     public void testEdit() throws Exception {
         // set up mocks, and args
-        int imageId = 7;
+        Long imageId = 7L;
 
         ServerImageForm serverImageForm = new ServerImageForm();
-        serverImageForm.setProviderAccountId(1);
+        serverImageForm.setProviderAccountId(1L);
         serverImageForm.setProviderImageId("id");
         serverImageForm.setDescription("desc");
         serverImageForm.setVersion("version");
