+++ account-management-security/src/main/java/org/duracloud/account/security/vote/AccountManagerAccessDecisionVoter.java
@@ -4,10 +4,10 @@
 package org.duracloud.account.security.vote;
 
 import org.aopalliance.intercept.MethodInvocation;
-import org.duracloud.account.common.domain.DuracloudUser;
-import org.duracloud.account.db.DuracloudRepoMgr;
+import org.duracloud.account.db.model.DuracloudUser;
+import org.duracloud.account.db.repo.DuracloudRepoMgr;
+import org.duracloud.account.db.util.AccountManagerService;
 import org.duracloud.account.security.domain.SecuredRule;
-import org.duracloud.account.util.AccountManagerService;
 import org.duracloud.common.error.DuraCloudRuntimeException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -60,12 +60,12 @@ public class AccountManagerAccessDecisionVoter extends BaseAccessDecisionVoter {
             decision = voteHasRole(role, userRoles);
 
         } else if (scope.equals(SecuredRule.Scope.SELF_ACCT)) {
-            int acctId = getIntArg(invocation.getArguments());
+            Long acctId = getIntArg(invocation.getArguments());
             decision = voteUserHasRoleOnAccount(user, role, acctId);
 
         } else if (scope.equals(SecuredRule.Scope.SELF_ID)) {
             if (voteHasRole(role, userRoles) == ACCESS_GRANTED) {
-                int userId = getIntArg(invocation.getArguments());
+                Long userId = getIntArg(invocation.getArguments());
                 decision = voteMyUserId(user, userId);
             }
 
@@ -79,11 +79,11 @@ public class AccountManagerAccessDecisionVoter extends BaseAccessDecisionVoter {
     }
 
 
-    private int getIntArg(Object[] arguments) {
+    private Long getIntArg(Object[] arguments) {
         if (arguments.length != 1) {
             log.error("Illegal number of args: " + arguments.length);
         }
-        return (Integer) arguments[0];
+        return (Long) arguments[0];
     }
 
     private int castVote(int decision, MethodInvocation invocation) {
